requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
fpdf==1.7.2
idna==3.4
imgkit==1.2.2
lxml==4.9.1
pdfkit==1.0.0
python-docx==0.8.11
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.12
